Removes drift dependency on toroidal current dir
Upon closer inspection/thought, the toroidal current induces a poloidal field. Reversing the direction of the toroidal current reverses the direction of the poloidal B-field. The impact of the drift flows is, then ExB = (Erad + Epol)x(Btor+Bpol) = Erad x Btor + Epol x Btor + Erad x Bpol + Epol x Bpol (-> 0). Thus, the only impact of reversing Itor on the drifts is a reversal of the toroidal shear flow. However, UEDGE assumes toroidicity, so this contribution is not considered.